Rover move and change direction methods

+++ Domain/Entities/Base.cs
@@ -3,7 +3,7 @@
     public class Base<TPrimaryKey>
     {
         public TPrimaryKey Id { get; set; }
-        public ushort PosX { get; set; }
-        public ushort PosY { get; set; }
+        public int PosX { get; set; }
+        public int PosY { get; set; }
     }
 }
